feat: update layout and components to use common directory structure, add MainBanner and StoryCard components for enhanced main page presentation
