+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import { Main } from '@krafton-soc/frames/Main';
+import Main from '@krafton-soc/features/Main';
 
 export default function HomePage() {
   return <Main />;
