+++ src/common/components/Header/LanguageSwitcher.tsx
@@ -2,8 +2,8 @@
 
 import React from 'react';
 import styled from '@emotion/styled';
-import { theme } from '@krafton-soc/styles/theme';
-import Text from '@krafton-soc/components/Text';
+import { theme } from '@krafton-soc/common/styles/theme';
+import Text from '@krafton-soc/common/components/Text';
 
 const LanguageToggle = styled.div`
   display: flex;
