+++ src/common/libs/RootStyleRegistry.tsx
