+++ src/common/styles/fonts.ts
@@ -4,27 +4,27 @@ import localFont from 'next/font/local';
 export const pretendard = localFont({
   src: [
     {
-      path: '../../public/fonts/Pretendard-Regular.woff2',
+      path: '../../../public/fonts/Pretendard-Regular.woff2',
       weight: '400',
       style: 'normal',
     },
     {
-      path: '../../public/fonts/Pretendard-Medium.woff2',
+      path: '../../../public/fonts/Pretendard-Medium.woff2',
       weight: '500',
       style: 'normal',
     },
     {
-      path: '../../public/fonts/Pretendard-Bold.woff2',
+      path: '../../../public/fonts/Pretendard-Bold.woff2',
       weight: '700',
       style: 'normal',
     },
     {
-      path: '../../public/fonts/Pretendard-ExtraBold.woff2',
+      path: '../../../public/fonts/Pretendard-ExtraBold.woff2',
       weight: '800',
       style: 'normal',
     },
     {
-      path: '../../public/fonts/Pretendard-Black.woff2',
+      path: '../../../public/fonts/Pretendard-Black.woff2',
       weight: '900',
       style: 'normal',
     },
