--- src/components/Header/index.ts
@@ -1,1 +1,0 @@
-export { default as Header } from './Header';
