+++ src/features/Main/components/MainBanner.tsx
@@ -3,8 +3,8 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import Image from 'next/image';
-import Text from '@krafton-soc/components/Text';
-import { theme } from '@krafton-soc/styles/theme';
+import Text from '@krafton-soc/common/components/Text';
+import { theme } from '@krafton-soc/common/styles/theme';
 
 const BannerContainer = styled.section`
   position: relative;
