--- src/frames/Main/index.ts
@@ -1,1 +1,0 @@
-export { default as Main } from './Main';
